nixos/networkd: Set Gateway= in [Network] section instead of [Route]
nixos/modules/tasks/network-interfaces-systemd.nix tries to read the values of `network.defaultGateway{,6}.address` and add them to `.network` files as gateways. When you specify both `defaultGateway` and `defaultGateway6`, it produces a `[Route]` section like this:

```
[Route]
Gateway=192.168.0.1
Gateway=dead:beef::1
GatewayOnLink=false
```

This doesn't work, because `Gateway=` can only be specified once in the `[Route]` section.

How this might have come about is that the man page for `systemd.network(5)` lists two different `Gateway=` options. The first one is in the `[Network]` block:

> `Gateway=`
>     The gateway address, which must be in the format described in `inet_pton(3)`. This is a short-hand for a `[Route]` section only containing a `Gateway` key. **This option may be specified more than once.**

The second one is in the `[Route]` block (which is what this code sets):

> `Gateway=`
>               Takes the gateway address or special value "_dhcp". If "_dhcp", then the gateway address provided by DHCP (or in the IPv6 case, provided by IPv6 RA) is used.

Notably, the latter doesn't say anything about being able to be specified multiple times. In addition, the top-level documentation for `[Route]` says "The `[Route]` section accepts the following keys. Specify several `[Route]` sections to configure several routes."

This commit changes the code to specify `Gateway=` in the `[Network]` section instead, where it can be specified multiple times, which may have been the original intention of the previous code. This means we lose the `GatewayOnLink=false` declaration, but as this is already the default, I think it's okay.

The other alternative would be to add a separate `[Route]` section for each gateway, but I don't see what the benefit of that would be.